Add test coverage for proto.Proto. Requires protoc.
