+++ go.mod
@@ -7,8 +7,9 @@ require (
 	github.com/bobg/errors v0.10.0
 	github.com/bobg/go-generics/v2 v2.1.1
 	github.com/bobg/tsdecls v0.1.0
+	github.com/bradleyjkemp/cupaloy/v2 v2.8.0
+	github.com/davecgh/go-spew v1.1.1
 	github.com/gibson042/canonicaljson-go v1.0.3
-	github.com/mattn/go-shellwords v1.0.12
 	github.com/mattn/go-sqlite3 v1.14.15
 	github.com/otiai10/copy v1.7.0
 	github.com/pressly/goose/v3 v3.6.1
@@ -19,10 +20,9 @@ require (
 
 require (
 	github.com/bobg/go-generics v1.5.0 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/fatih/camelcase v1.0.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/exp v0.0.0-20230213192124-5e25df0256eb // indirect
-	golang.org/x/sync v0.0.0-20220819030929-7fc1605a5dde // indirect
 	golang.org/x/sys v0.1.0 // indirect
 )
