+++ golang/go_test.go
@@ -40,9 +40,8 @@ func TestGo(t *testing.T) {
 		"../hash_test.go",
 		"../main.go",
 		"../main_test.go",
-		"../proto/deps.go",
-		"../proto/deps_test.go",
-		"../proto/rule.go",
+		"../proto/proto.go",
+		"../proto/proto_test.go",
 		"../register.go",
 		"../register_test.go",
 		"../runner.go",
