+++ proto/testdata/foo.proto
@@ -1,2 +1,8 @@
+syntax = "proto3";
+
 import "x/bar.proto";
 import "y/baz.proto"; // does not exist
+
+message M {
+  string x = 1;
+}
