+++ proto/testdata/foo2.proto
@@ -1,0 +1,8 @@
+syntax = "proto3";
+
+import "x/bar2.proto";
+
+message M {
+  string x = 1;
+  M2     y = 2;
+}
