+++ proto/testdata/x/bar.proto
@@ -1,1 +1,3 @@
+syntax = "proto3";
+
 import public "x/plugh.proto";
